requirements.txt
requests==2.31.0
pandas==2.0.3
numpy==1.24.3
Flask==3.0.0
APScheduler==3.10.1
python-socketio==5.9.0
dash==2.14.0
dash-bootstrap-components==1.5.0
psycopg2-binary==2.9.9